+++ src/app/Playground/VetleSamples/ViewCellTest.xaml.cs
@@ -4,6 +4,8 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Input;
+using DIPS.Mobile.UI.Components.ChipGroup;
+using DIPS.Mobile.UI.Components.Pickers.ItemPicker;
 
 namespace Playground.VetleSamples;
 
@@ -36,4 +38,23 @@ public partial class ViewCellTest : ContentView
         set => SetValue(TestBoolProperty, value);
     }
 
+    private void ItemPicker_OnDidSelectItem(object sender, object e)
+    {
+        if(BindingContext is not TimePlanningViewModel timePlanningViewModel)
+            return;
+
+        if(sender is not ItemPicker itemPicker)
+            return;
+
+        
+    }
+
+    private async void DateAndRoundPicker_OnOnSelectedItemsChanged(object sender, ChipGroupEventArgs e)
+    {
+        if(BindingContext is not TimePlanningViewModel timePlanningViewModel)
+            return;
+        
+        await Task.Delay(1000);
+        timePlanningViewModel.Update();
+    }
 }