+++ src/library/DIPS.Mobile.UI/Components/Images/Image/dotnet/ImageHandler.cs
@@ -1,13 +1,3 @@
 namespace DIPS.Mobile.UI.Components.Images.Image;
 
-public partial class ImageHandler
-{
-    private static partial void TrySetTintColor(ImageHandler handler, Image image)
-    {
-    }
-
-    private static partial void AppendToPropertyMapper()
-    {
-    }
-
-}
+public partial class ImageHandler { }