+++ src/library/DIPS.Mobile.UI/Components/Images/ImageButton/Android/ImageButtonHandler.cs
@@ -11,8 +11,6 @@ namespace DIPS.Mobile.UI.Components.Images.ImageButton;
 
 public partial class ImageButtonHandler
 {
-
-    private partial void AppendPropertyMapper() { }
     protected override void ConnectHandler(ShapeableImageView platformView)
     {
         base.ConnectHandler(platformView);
@@ -28,27 +26,14 @@ public partial class ImageButtonHandler
 
     private static partial void TrySetTintColor(ImageButtonHandler handler, ImageButton imageButton)
     {
-        if (handler.PlatformView is null)
-        {
-            DUILogService.LogError<ImageButtonHandler>("PlatformView is null, this should not happen, most likely the issue is that the Content is rendered and then the handler is instantly disconnected");
-            return;
-        }
-            
-        handler.PlatformView?.SetColorFilter(imageButton.TintColor.ToPlatform());
+        handler.PlatformView.SetColorFilter(imageButton.TintColor.ToPlatform());
     }
 
     private static async partial void MapAdditionalHitBoxSize(ImageButtonHandler handler, ImageButton imageButton)
     {
         await Task.Delay(1);
 
-        if (handler.PlatformView is null)
-        {
-            DUILogService.LogError<ImageButtonHandler>("PlatformView is null, this should not happen, most likely the issue is that the Content is rendered and then the handler is instantly disconnected");
-            return;
-        }
-        
-        handler.PlatformView?.SetAdditionalHitBoxSize(imageButton, imageButton.AdditionalHitBoxSize, handler.MauiContext!);
-        
+        handler.PlatformView.SetAdditionalHitBoxSize(imageButton, imageButton.AdditionalHitBoxSize, handler.MauiContext!);
     }
 }