+++ src/library/DIPS.Mobile.UI/Components/Images/ImageButton/dotnet/ImageButtonHandler.cs
@@ -2,17 +2,7 @@ namespace DIPS.Mobile.UI.Components.Images.ImageButton;
 
 public partial class ImageButtonHandler
 {
-    private static partial void TrySetTintColor(ImageButtonHandler handler, ImageButton imageButton)
-    {
-    }
-
     private static partial void MapAdditionalHitBoxSize(ImageButtonHandler handler, ImageButton imageButton)
     {
     }
-
-    private partial void AppendPropertyMapper()
-    {
-        
-    }
-    
 }